requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2

chromadb==0.5.5

openai==1.40.6
httpx==0.27.2

python-multipart==0.0.9
